Renaming library and profiles
