+++ examples/connext_dds/network_capture/01_shared_memory_and_udp/c++03/USER_QOS_PROFILES.xml
@@ -11,8 +11,8 @@
  -->
 <dds xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
      xsi:noNamespaceSchemaLocation="http://community.rti.com/schema/6.1.0/rti_dds_qos_profiles.xsd">
-    <qos_library name="Library">
-      <qos_profile name="Profile" base_name="BuiltinQosLibExp::Generic.StrictReliable" is_default_qos="true">
+    <qos_library name="network_capture_Library">
+      <qos_profile name="network_capture_Profile" base_name="BuiltinQosLibExp::Generic.StrictReliable" is_default_qos="true">
         <domain_participant_qos>
           <!--
           This property breaks compatibility with 6.0.0 but without it
