+++ examples/connext_dds/network_capture/01_shared_memory_and_udp/cs/network_capture.idl
@@ -9,6 +9,6 @@
  * not be liable for any incidental or consequential damages arising out of the
  * use or inability to use the software.
  */
-struct example0 {
+struct NetworkCapture {
     string<128> msg;
 };
