+++ examples/connext_dds/network_capture/02_tcp/c/USER_QOS_PROFILES.xml
@@ -12,8 +12,8 @@
 <dds xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
      xsi:noNamespaceSchemaLocation="http://community.rti.com/schema/6.1.0/rti_dds_qos_profiles.xsd"
      version="6.1.0">
-    <qos_library name="Library">
-      <qos_profile name="Default" base_name="BuiltinQosLibExp::Generic.StrictReliable" is_default_qos="true">
+    <qos_library name="network_capture_Library">
+      <qos_profile name="network_capture_Profile" base_name="BuiltinQosLibExp::Generic.StrictReliable" is_default_qos="true">
         <domain_participant_qos>
           <property>
             <value>
@@ -45,7 +45,7 @@
           </property>
         </domain_participant_qos>
       </qos_profile>
-      <qos_profile name="LANdefault" base_name="Default">
+      <qos_profile name="LANdefault" base_name="network_capture_Profile">
         <domain_participant_qos>
           <property>
             <value>
@@ -94,7 +94,7 @@
           </discovery>
         </domain_participant_qos>
       </qos_profile>
-      <qos_profile name="WANdefault" base_name="Default">
+      <qos_profile name="WANdefault" base_name="network_capture_Profile">
         <domain_participant_qos>
           <property>
             <value>
@@ -185,7 +185,7 @@
         </property>
       </domain_participant_qos>
     </qos_profile>
-    <qos_profile name="TLSdefault" base_name="Default">
+    <qos_profile name="TLSdefault" base_name="network_capture_Profile">
       <domain_participant_qos>
         <property>
           <value>
