+++ examples/connext_dds/network_capture/02_tcp/c/network_capture_publisher.c
@@ -117,7 +117,7 @@ int publisher_main(int domainId, int sample_count, const char *profile)
     participant = DDS_DomainParticipantFactory_create_participant_with_profile(
             DDS_TheParticipantFactory,
             domainId,
-            "Library",
+            "network_capture_Library",
             profile,
             NULL, /* listener */
             DDS_STATUS_MASK_NONE);
