+++ examples/connext_dds/network_capture/04_advanced_api/c/USER_QOS_PROFILES.xml
@@ -11,8 +11,8 @@
  -->
 <dds xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
      xsi:noNamespaceSchemaLocation="http://community.rti.com/schema/6.1.0/rti_dds_qos_profiles.xsd">
-    <qos_library name="SecurityExampleProfiles">
-        <qos_profile name="A" base_name="BuiltinQosLib::Generic.Security" is_default_qos="true">
+    <qos_library name="network_capture_Library">
+        <qos_profile name="network_capture_ProfileA" base_name="BuiltinQosLib::Generic.Security" is_default_qos="true">
           <domain_participant_qos>
             <property>
               <value>
@@ -52,7 +52,7 @@
             </property>
           </domain_participant_qos>
         </qos_profile>
-      <qos_profile name="B" base_name="A">
+      <qos_profile name="network_capture_ProfileB" base_name="network_capture_ProfileA">
         <domain_participant_qos>
           <property>
             <value>
